Moves common buttons to global components
